GBZ80: Update the calling convention and clean up register defs.
8-bit values are returned in A, 16-bit values
in HL.

For parameters, the calling convention tries to
allocate A/HL, E/DE, C/BC, and then defers to the
stack.

Callee-saved regs are BC & DE for 0-1 arguments,
BC for 2 arguments, and none for 3+.

Also clean up allocation order and register costs.
The allocation order is now A, L, H, E, D, C, B for
8-bit regs to hopefully encourage better allocation
of low regs, and HL, DE, BC for 16-bit regs.