GBZ80: Add PostPEI pass.
This will probably do something.

It might be more efficient to do
push-pop optimization before PEI
since we can't compress the stack
after PEI is over.

One idea is to do the analysis of
push-pop in PostRA, then use the
information during PEI instead of
after? But that might not let us
do what we want.